requirements.txt
pathlib
yaml
json
requests
pygithub
argparse
termcolor